BUG FIX: Syncing currentTimestamp and layer updates
Debounces setting of currentTimestamp instead of updating the layer
in the layer controller. Debouncing the layer updating allowed for
the currentTime and the layer to become out of sync. Since everything
is synced to update on changes to the currentTimstamp, debouncing
updates to the currentTimestamp ensures everything remains in sync.